fix: Rename dead constant; Improvememnts
