+++ plugin/musixmatch/2_desktop_e2e_test.go
@@ -95,16 +95,16 @@ func mobileFirstMacroFixture(t *testing.T) []byte {
 	trackBody, _ := json.Marshal(map[string]any{"track": map[string]string{"track_name": "Birds of a Feather", "artist_name": "Billie Eilish", "album_name": ""}})
 	subtitleBody, _ := json.Marshal(map[string]any{"subtitle_list": []any{map[string]any{"subtitle": map[string]string{"subtitle_body": "[00:01.00]Birds of a feather\n[00:02.00]We should stick together\n", "subtitle_language": "en"}}}})
 	track := desktopResponse{}
-	track.Message.Header.StatusCode = statusCodePISuccess
+	track.Message.Header.StatusCode = utils.HTTPStatusOK
 	track.Message.Body = trackBody
 	subtitle := desktopResponse{}
-	subtitle.Message.Header.StatusCode = statusCodePISuccess
+	subtitle.Message.Header.StatusCode = utils.HTTPStatusOK
 	subtitle.Message.Body = subtitleBody
 	macro.MacroCalls["matcher.track.get"] = track
 	macro.MacroCalls["track.subtitles.get"] = subtitle
 	body, _ := json.Marshal(macro)
 	outer := desktopResponse{}
-	outer.Message.Header.StatusCode = statusCodePISuccess
+	outer.Message.Header.StatusCode = utils.HTTPStatusOK
 	outer.Message.Body = body
 	out, _ := json.Marshal(outer)
 	return out
@@ -122,7 +122,7 @@ func withMatchedTrackFixture(t *testing.T, macroBody []byte, title, artist, albu
 	}
 	trackBody, _ := json.Marshal(map[string]any{"track": map[string]string{"track_name": title, "artist_name": artist, "album_name": album}})
 	call := desktopResponse{}
-	call.Message.Header.StatusCode = statusCodePISuccess
+	call.Message.Header.StatusCode = utils.HTTPStatusOK
 	call.Message.Body = trackBody
 	if macro.MacroCalls == nil {
 		macro.MacroCalls = map[string]desktopResponse{}
@@ -152,7 +152,7 @@ func mustGetDesktopMacroBody(t *testing.T, tokenBody []byte) []byte {
 	if err := json.Unmarshal(tokenBody, &tokenResp); err != nil {
 		t.Fatalf("failed to parse desktop token fixture: %v", err)
 	}
-	if tokenResp.Message.Header.StatusCode != statusCodePISuccess {
+	if tokenResp.Message.Header.StatusCode != utils.HTTPStatusOK {
 		t.Fatalf("desktop token fixture returned status %d: %s", tokenResp.Message.Header.StatusCode, tokenBody)
 	}
 	if len(tokenResp.Message.Body) == 0 {
