+++ plugin/musixmatch/4_website_scraper.go
@@ -3,54 +3,12 @@ package musixmatch
 import (
 	"encoding/json"
 	"fmt"
-	"regexp"
 	"strings"
 
 	"github.com/Myzel394/navidrome-musixmatch-plugin/plugin/utils"
 	"github.com/navidrome/navidrome/plugins/pdk/go/lyrics"
 	"github.com/navidrome/navidrome/plugins/pdk/go/pdk"
 )
-
-var nextDataRe = regexp.MustCompile(`(?s)<script id="__NEXT_DATA__" type="application/json">(.*?)</script>`)
-
-type subtitleLine struct {
-	Text string `json:"text"`
-	Time struct {
-		Total      float64 `json:"total"`
-		Minutes    int     `json:"minutes"`
-		Seconds    int     `json:"seconds"`
-		Hundredths int     `json:"hundredths"`
-	} `json:"time"`
-	Type string `json:"type"`
-}
-
-type sectionLine struct {
-	Title string         `json:"title"`
-	Lines []subtitleLine `json:"lines"`
-}
-
-type nextDataResponse struct {
-	Props struct {
-		PageProps struct {
-			Data struct {
-				TrackInfo struct {
-					Data struct {
-						Track struct {
-							AlbumName  string `json:"albumName"`
-							ArtistName string `json:"artistName"`
-						} `json:"track"`
-						Lyrics struct {
-							Body     string `json:"body"`
-							Language string `json:"language"`
-						} `json:"lyrics"`
-						TrackStructureList []sectionLine  `json:"trackStructureList"`
-						Subtitle           []subtitleLine `json:"subtitle"`
-					} `json:"data"`
-				} `json:"trackInfo"`
-			} `json:"data"`
-		} `json:"pageProps"`
-	} `json:"props"`
-}
 
 func fetchLyricsViaWebsiteScraping(input lyrics.GetLyricsRequest) (lyrics.GetLyricsResponse, error, *utils.LookupFailure, *utils.LookupSuccess) {
 	// Fallback, scrape website
@@ -123,7 +81,7 @@ func scrapeWebsiteLyricsForTrack(track *Song, input lyrics.GetLyricsRequest) (ly
 	}
 	utils.LogInfof("website lyrics page: response received body_bytes=%d", len(body))
 
-	matches := nextDataRe.FindSubmatch(body)
+	matches := WEBSITE_NEXT_DATA_REGEX.FindSubmatch(body)
 	if len(matches) < 2 {
 		utils.LogErrorf("website lyrics page: Next.js data not found body_bytes=%d", len(body))
 		pdk.Log(pdk.LogDebug, fmt.Sprintf("website lyrics page: response body=%s", string(body)))
